requirements.txt
ccxt
openai
requests
python-dotenv
